vibration/api-is-present.html

<!DOCTYPE html>
<html xmlns='http://www.w3.org/1999/xhtml' lang='en'>
  <head>
    <meta charset='utf-8'/>
    <title>Vibration API: test that the vibrate() method is present (with or without vendor prefix)</title>
    <link rel='author' title='Robin Berjon' href='mailto:robin@berjon.com'/>
    <link rel='help' href='http://www.w3.org/TR/vibration/#methods'/>
    <meta name='flags' content='vendor-prefix, dom'/>
    <meta name='assert' content='Check that the vibrate() method is present.'/>
    <!-- 
      When you review this test, please add
      <link rel='reviewer' title='Your Name' href='mailto:your-email-address'/>
    -->
    <link rel='stylesheet' href='/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      This test checks for the presence of the <code>vibrate()</code> method, taking
      vendor prefixes into account.
    </p>
    <div id='log'></div>
    <script src='/resources/testharness.js'></script>
    <!-- When this test gets approved, remove a "../" below -->
    <script src='support/feature-detection.js'></script>
    <script>
      test(function () {
          assert_true(undefined !== BrowserHasFeature(navigator, "vibrate"), "navigator.vibrate exists");
      }, "vibrate() is present on navigator");
    </script>
  </body>
</html>
